requirements.txt
alembic==1.0.5
awesome-slugify==1.6.5
bcrypt==3.1.5
cffi==1.11.5
Click==7.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Migrate==2.3.1
Flask-MySQL==1.4.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
mysql-connector-python==8.0.14
Pillow==5.4.0
protobuf==3.6.1
pycparser==2.19
PyMySQL==0.9.3
python-dateutil==2.7.5
python-editor==1.0.3
regex==2019.1.24
six==1.12.0
SQLAlchemy==1.2.15
Unidecode==0.4.21
Werkzeug==0.14.1
WTForms==2.2.1
